+++ Changelog.md
@@ -10,6 +10,7 @@
 - API: IEMPAffected has been added. It can be used to shut down select SecurityCraft blocks and entities when they're within range of an EMP blast
 - API: IEMPAffectedTE has been added as a default implementation for tile entities which want to implement IEMPAffected
 - Fix: The default display name of some blocks is always "Customize"
+- Fix: Crash with disguised blocks from before the state selector existed
 
 --------------------------Changelog for v1.9.2 of SecurityCraft--------------------------